Persistent internal Reader buffer added to speed up Row reads.
The buffer is implemented using array doubling. After a read, it is
zeroes out and the size is unchanged. This greatly reduces the number of
allocations when reading large files.

It should be noted that the string casts performed on any byte buffer
slices will copy and protect the data from being cleared from the
process above.